Updated JSHint configuration
This change removes deprecated JSHint options and adds a couple
new ones. Many of the validation options that were available have
been removed in favor of using JSCS in addition to JSHint.

Using ES5 "esversion" because the package.json file still says that
this package is compatible with Node.js version 0.8.